requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pydantic
email-validator
python-multipart
PyPDF2
SpeechRecognition
gTTS
google-generativeai
